Feat: 탱크 상속구조 추가
1. 이동, 회전이 구현된 일반 탱크클래스와 키맵핑이 이루어지는 플레이어 탱크클래스를
   따로 분리
